move gate class
